+++ src/renderer/components/dialogs/CreateChat.js
@@ -7,22 +7,24 @@ import C from 'deltachat-node/constants'
 import { callDcMethodAsync } from '../../ipc'
 import ScreenContext from '../../contexts/ScreenContext'
 import { useContacts, ContactList2, PseudoContactListItem } from '../helpers/ContactList'
-import { isValidEmail } from '../helpers/Contact'
 import { DeltaButtonPrimary } from '../helpers/SmallDialog'
 import { DeltaDialogBase, DeltaDialogCloseButton, GoBackDialogHeader } from '../helpers/DeltaDialog'
 import {
-  GroupSettingsContainer,
-  GroupSeperator,
-  GroupMemberContactListWrapper,
   CreateChatSearchInput,
   CreateChatContactListWrapper,
-  GroupImage,
-  GroupNameInput,
-  GroupMemberSearchInput,
   PseudoContactListItemNoSearchResults,
   PseudoContactListItemShowQrCode,
   PseudoContactListItemAddMember
 } from './CreateChat-Styles'
+
+import {
+  GroupSettingsContainer,
+  GroupSeperator,
+  GroupMemberContactListWrapper,
+  GroupImage,
+  GroupNameInput,
+  GroupMemberSearchInput
+} from './Group-Styles'
 
 export default function CreateChat (props) {
   const { isOpen, onClose } = props
@@ -479,4 +481,24 @@ export function CreateVerifiedGroupInner (props) {
       }
     </>
   )
+}
+
+function isValidEmail (email) {
+  // empty string is not allowed
+  if (email === '') return false
+  const parts = email.split('@')
+  // missing @ character or more than one @ character
+  if (parts.length !== 2) return false
+  const [local, domain] = parts
+  // empty string is not valid for local part
+  if (local === '') return false
+  // domain is too short
+  if (domain.length <= 3) return false
+  const dot = domain.indexOf('.')
+  // invalid domain without a dot
+  if (dot === -1) return false
+  // invalid domain if dot is (second) last character
+  if (dot >= domain.length - 2) return false
+
+  return true
 }
